[ADD] Feature fields of equality
It helps to not duplicate dest records that already exists before
synchronization. It checks equality in dest using the fields set at
fields_of_equality.
For example, to check account.journal, we can use code and name to check
if it exists in dest. So it will not be duplicated and avoid constraint
error.