[FIX] partner_credit_limit: account type
We can't simply use `name` of model `account.account.type` to search for
related move lines. Because firstly `name` is not fixed (users can
change name to anything they want), second `name` field is translatable,
so if language is changed to other then English, and there is a
translation, it will return incorrect results, because search will never
find any related moves anymore.

Instead we use `type` of model `account.account.type` which is
consistent and can't be changed that simply by user.