[FIX] partner_credit_limit: api.one -> api.multi
When we use `api.one` decorator and return result, it will be mangled in
a list, so if we would expect for example `False` to be returned, it
would return `[False]` instead which could potentially lead to incorrect
check.

In this case, true/false is not checked, but still if we need to make
sure method is called from only one record, we have to use
`self.ensure_one` instead.